Adding more notes and links, working on simpler names and more axis types

+++ include/boost/histogram/python/storage.hpp
@@ -2,7 +2,8 @@
 
 #include <boost/histogram.hpp>
 #include <vector>
+#include <cstdint>
 
 namespace bh = boost::histogram;
 
-using vector_int_storage = bh::storage_adaptor<std::vector<int>>;
+using vector_int_storage = bh::storage_adaptor<std::vector<uint64_t>>;
